Allow to ignore paths with regular expression
Right now we need to exclude all files in a particular directory or
files with particular name we will need to pass a full list into an
ignore method. In order to simplify this workflow we could add ability
to ignore paths based on regular expression that Dir.glob can recognize.