delete autoload paths in the registry when files are loaded
Once the file has been visited, we do not need its entry in the registry
any more. This saves some memory.

This is not so easy for directories, because Ruby does not serialize the
require calls and our Kernel#require wrapper could invoke the original
method for a directory, failing.