removes the docs for preloading
I no longer have a use case for preloading since the recommendation for
STIs in Rails shifted to a different technique.  Also, this feature is
based on require, which is something we are avoiding. In Zeitwerk, you
force loading things via const_get, not require.

Unless we find a use case for preloading, this feature might be gone in
some future release.