simplify .travis.yml
For some reason, Travis hanged when building for ruby-head in the
last, default distro. The log stopped when importing gems into a
gemset, for example see

    https://travis-ci.com/fxn/zeitwerk/jobs/244977328

So not related to Zeitwerk itself, seems a problem with provisioning.

Support suggested to remove Gemfile.lock from the repository instead
of doing so in a before_install hook. I do not know why does that fix
the problem, but it does.