Extract configuration to its own module
With time, lib/zeitwerk/loader.rb has gotten too many lines for my taste.

While the majority of the public API of Zeitwerk::Loader for end users is
configuration API, the implementation has configuration-related code, and
loading-related code, so to speak. This patch separates them for ease
understanding and internal code organization.

I played with the idea of an internal configuration class and wrote some PoC,
but added unnecessary complexity. For me, this is a better trade-off.