Small refactor in raise_if_conflicting_directory
We only need to compute the slashed dir once. Then, we use a similar variable
name for the slashed root dir for consistency.

As a bonus, I think the comparisons are arguably easier to understand, since you
have the existence of the trailing slashes right there in the variable names.