Mark helpers as private individually
Sometimes, I am in this file and wonder about the formal visibility. Then I have
to scan upwards until I find the top private call. Should not make me think.

This is part of a trend in the project in which everything that is not public is
obviously so just by casually reading the source code.