Let push_dir raise if given an anonymous root namespace
Anonymous root namespaces are not supported, Zeitwerk assumes classes and
modules have regular names in a few places.

However, if you pass an anonymous root namespace right now the library crashes
in some random internal spot. It is better to tell the user at the boundary and
in a controlled way. Also, this allows us to work internally knowing that root
namespaces in `roots` have a name for sure.