Reset module state on `Zeitwerk::NameError`
If an autoload is triggered, the file is loaded successfully, but the expected
constant does not get defined, Ruby resets the state of the module. In
particular, `autoload?` returns `nil` for that constant name, and `constants`
does not include the constant name (starting with Ruby 3.1).

Zeitwerk is more strict, not defining the expected constant is an error
condition and the loader raises `Zeitwerk::NameError`. But this happens during
the `require` call and the exception prevents Ruby from doing that cleanup.

This does not matter if the exception aborts the program. However, if the user
rescues the error, the module is left in a state that makes more sense.