Delete comment in Kernel#require decoration
That was my reasoning when I wrote that code. However, I've noticed a
few references to zeitwerk_original_require in a GitHub search.
Moreover, I wonder whether altering the ancestor chain of Kernel or
Object could be problematic in practice.

All in all, even if we drop some day support for Ruby < 3.0, I am not
sure refactoring this code is worth it.