Reword Zeitwerk::NameError message again
I am sleeping on this one and trying different options.

The previous revision lacked symmetry for my taste, because if you say
"the top-level constant X", when there is a namespace "M::N", why is
that "M" not qualified as top-level?

The answer is that an error message is not a constant lookup, it is by
convention relative to the top-level namespace. But then, the messages
for simple and compound constant paths lack symmetry.

This alternative is more coherent, and also matches the way you think
about project layouts. You define constants in namespaces represented by
directories.

Let's try this one now.