Skipping `autoload?` call by manually defining the edge cases
In our journey of improving our Rails application bootstrap we have detected that calling the method `autoload?` for each `autoload` registration is very costly. In our case it takes >3s.

In this commit we try to manually define the edge cases that we detected in our codebase so we can benefit of calling this logic only when it's strictly needed.

The idea would be to evolve this, so it could be configured externally and everybody can benefit of disabling this costly call.