Manually defining inception namespaces to skip `autoload?` call
In our journey of improving our Rails application bootstrap we have
detected that calling the method `Module.autoload?` for each `autoload`
registration is very costly.
In our case it takes >3s. (see attached flamegraphs)

As the comment in the implementation points out, this call is there to
support an edge case on how certain gems autoload themselves. [More info here](https://github.com/fxn/zeitwerk/blob/main/lib/zeitwerk/registry.rb#L28)
However this approach produces [a
call](https://github.com/fxn/zeitwerk/blob/main/lib/zeitwerk/loader.rb#L540)
to `Module.autoload?` for every single constant in our huge project. As
it can be seen in the attached flamegraph. This is taking for our
backend more than 3 seconds just to support few gems with this special
case.

In this PR we propose to enable a mechanism on which any Rails
application can manually define the namespaces to be incepted and avoid
the costly call for the rest of the constants. This resulted in a boot
reduction of more than 3 seconds in our case. We only had to define 3
namespaces to be incepted in order to make our backend run properly.

As a trade-off I understand that this will cause confusion to developers
including gems needing the inception. So this PR proposes to keep
backwards compatibility if you don't manually set `manual_incepted_namespaces`.

TODO:
- [ ] Add meaningful tests