Define Zeitwerk::Cref#{to_s,hash,eql?,==}
This refactor makes the code a bit simpler and aligned with the introduction of
Zeitwerk::Cref.

The performance of `unloadable_cpaths` and `unloadable_cpath?` is worse now, but
those are **undocumented** methods (therefore lib-private) that were written
only to help transtion from the classic autoloader API in Rails 6.

That API was rarely used back then, it was also undocumented, and its use cases
were casual introspections when you wanted to debug something.

Rails 7.0 removed classic and, with that, that API and usage of these methods.

These methods should be deleted at some point.