Create a proper registry for explicit namespaces
When I wrote this feature, I initially defined a module

    Zeitwerk::Registry::ExplicitNamespaces

because this is actually a registry, as the documentation acknowledges:

> This module is essentially a registry for explicit namespaces.

However, Module#const_added is a hot path and I fine-tuned the hell of it. I
moved the module up to the Zeitwerk namespace because I measured having one
extra segment in the constant path had a measurable impact.

Inspired by a refactor I am doing in another branch, I have revised this.

Now, we have the natural constant path Zeitwerk::Registry::ExplicitNamespaces,
but the module also defines the top-level constant:

    ZEITWERK_REGISTRY_EXPLICIT_NAMESPACES

for Module#const_added.

I think being top-level should be fine due to the prefix.