Refactors the registry for explicit namespaces
The design of the registry is in flux, it grew and I felt it needed some
structure.

The first tentative steps have been directed towards splitting the registry into
several registries for clarity. So now, for example, we have a separate registry
for explicit namespaces.

That addresses part of the problem. But that does not feel just right to me, it
is too decoupled from Zeitwerk::Registry, and you can see that in the lack of
encapsulation when unregistering a loader, and kind of a non-uniform API.

My current intuition is that Zeitwerk::Registry should provide the API loaders
interact with, and that be split into inner registries.

Let's see where this approach takes us to.