Use #: for RBS annotations
I have annotated signatures since the first version of Zeitwerk. For me, they
are an essential aid to understand and use the code, both public and private.

In the case of Zeitwerk, the API is described in English in the README. These
signatures are mostly internal documentation for readers of the code.

You need to know what a method accepts, returns, yields, and what may raise. And
that is the case regardless of the dynamic or static nature of your type system.

At the beginning I used my own notation, then adopted RBS, and now a consensus
towards using #: as a marker seems to be taking shape.

This pull request to Sorbet

    https://github.com/sorbet/sorbet/pull/8470

inspired me to join this trend.