Allow for root_file customization
This allows for specification of where the `for_gem` and
`for_gem_extension` root file is located instead of relying
souly on `caller_locations` to make that determination.
This is useful for allowing custom loader wrapper classes
that invoke the loader from places other than the project root.

Further context can be found here:

* https://github.com/fxn/zeitwerk/issues/314
* https://github.com/fxn/zeitwerk/issues/314#issuecomment-2887058756